# attestationutil.AttestingIndices: Minor improvement on indices array allocation