--- a/shared/attestationutil/attestation_utils.go
+++ b/shared/attestationutil/attestation_utils.go
@@ -67,7 +67,7 @@ func ConvertToIndexed(ctx context.Context, attestation *ethpb.Attestation, commi
 //    committee = get_beacon_committee(state, data.slot, data.index)
 //    return set(index for i, index in enumerate(committee) if bits[i])
 func AttestingIndices(bf bitfield.Bitfield, committee []uint64) []uint64 {
-	indices := make([]uint64, 0, len(committee))
+	indices := make([]uint64, 0, bf.Count())
 	for _, idx := range bf.BitIndices() {
 		if idx < len(committee) {
 			indices = append(indices, committee[idx])