**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

This PR properly sets the size of the indices array. This improvement is most effective on partial committee bitfields. The previous implementation would allocation enough space for the whole committee when only a subset of indices had actually attested. 


Before

```
goos: linux
goarch: amd64
pkg: github.com/prysmaticlabs/prysm/shared/attestationutil
BenchmarkAttestingIndices_PartialCommittee-8     4581206               256 ns/op             544 B/op          3 allocs/op
PASS
ok      github.com/prysmaticlabs/prysm/shared/attestationutil   2.916s

```

After

```
goos: linux
goarch: amd64
pkg: github.com/prysmaticlabs/prysm/shared/attestationutil
BenchmarkAttestingIndices_PartialCommittee-8     4803296               251 ns/op             480 B/op          3 allocs/op
PASS
ok      github.com/prysmaticlabs/prysm/shared/attestationutil   2.889s
```

**Which issues(s) does this PR fix?**

N/A - General memory usage issues

**Other notes for review**

This PR is a one line change in production code. I tweaked benchmark to have less indices to amplify the performance difference in the results.